refactor(persistence): split event store helpers [OPE-421]
